# docs-api/openapi.yaml
openapi: 3.0.0
info:
  title: Vol-Web
  version: 1.0.0
servers:
  - url: http://localhost:8080/api
    description: Local Server
security:
  - express-session: []
paths:
...
  /projects:
    get:
      summary: List of all user projects
      operationId: getProjects
      tags:
        - Project
      responses:
        "200":
          description: An array of user's projects.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"

    post:
      summary: Create a new project
      operationId: createProject
      tags:
        - Project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - name
              - description
              properties:
                name:
                  type: string
                description:
                  type: string
      responses:
        '201':
          description: Created project.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
  /project/{idProject}:
    get:
      summary: Detailed info about a project.
      operationId: getProjectById
      tags:
        - Project
      parameters:
        - $ref: '#/components/parameters/idProject'
      responses:
        '200':
          description: Requested project.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        '404':
          description: Project not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      summary: Deletes a project.
      operationId: deleteProject
      tags:
        - Project
      parameters:
        - $ref: '#/components/parameters/idProject'
      responses:
        '204':
          description: Project successfuly deleted.
        '404':
          description: Project not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /project/{idProject}/volumes:
    post:
      summary: Create a new volume
      operationId: createVolume
      tags:
        - Volume
      parameters:
        - $ref: '#/components/parameters/idProject'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - name
              - description
              properties:
                name:
                  type: string
                description:
                  type: string
      responses:
        '201':
          description: Created Volume.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Volume'
...
  /project/{idProject}/volume/{idVolume}/raw-data/upload:
    put:
      summary: Upload raw volume data
      operationId: uploadRawData
      tags:
        - Volume
      parameters:
        - $ref: '#/components/parameters/idProject'
        - $ref: '#/components/parameters/idVolume'
      requestBody:
        required: true
        content:
          multipart/form-data:
              schema:
                type: object
                properties:
                  files:
                    type: array
                    items:
                      type: string
                      format: binary
                    description: Multiple files of type .zip, .raw, or .json
      responses:
        '200':
          description: Updated raw data volume.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RawData"
...
  /project/{idProject}/volume/{idVolume}/add-sparse-label-volume:
    put:
      summary: Add sparse label volume
      operationId: addSparseLabelVolume
      tags:
        - Volume
      parameters:
        - $ref: '#/components/parameters/idProject'
        - $ref: '#/components/parameters/idVolume'
      responses:
        '201':
          description: Created sparse label volume.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SparseLabelVolumeData"
  /project/{idProject}/volume/{idVolume}/add-pseudo-label-volume:
    put:
      summary: Add pseudo label volume
      operationId: addPseudoLabelVolume
      tags:
        - Volume
      parameters:
        - $ref: '#/components/parameters/idProject'
        - $ref: '#/components/parameters/idVolume'
      responses:
        '201':
          description: Created pseudo label volume.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PseudoLabelVolumeData"
  /project/{idProject}/volume/{idVolume}/add-annotations:
    put:
      summary: Add annotations and convert them to sparse label volume
      operationId: addAnnotations
      tags:
        - Volume
      parameters:
        - $ref: '#/components/parameters/idProject'
        - $ref: '#/components/parameters/idVolume'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                annotations:
                  type: object
                  description: A JSON object containing annotation data.

      responses:
        '201':
          description: Created sparse label volume.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SparseLabelVolumeData"
  /project/{idProject}/volume/{idVolume}/volumeData/:type/{idVolumeData}/upload-files:
    put:
      summary: Upload volume data files
      operationId: uploadFiles
      tags:
        - Volume Data
      parameters:
        - $ref: '#/components/parameters/idProject'
        - $ref: '#/components/parameters/idVolume'
        - $ref: '#/components/parameters/idVolumeData'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                files:
                  type: array
                  items:
                    type: string
                    format: binary
                  description: Multiple files of type .zip, .raw, or .json

      responses:
        '200':
          description: Updated raw data.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/RawData"
                  - $ref: "#/components/schemas/SparseLabelVolumeData"
                  - $ref: "#/components/schemas/PseudoLabelVolumeData"
...
  /project/{idProject}/models:
    post:
      summary: Create a new model
      operationId: createModel
      tags:
        - Model
      parameters:
        - $ref: '#/components/parameters/idProject'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - name
              - description
              properties:
                name:
                  type: string
                description:
                  type: string
      responses:
        '201':
          description: Created Model.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Model'
...
  /project/{idProject}/model/{idModel}/checkpoint/{idCheckpoint}/download:
    get:
      summary: Download mrc file in volume data 
      operationId: downloadCheckpoint
      tags:
        - Checkpoint
      parameters:
        - $ref: '#/components/parameters/idProject'
        - $ref: '#/components/parameters/idModel'
        - $ref: '#/components/parameters/idCheckpoint'
      responses:
        '200':
          description: ZIP file containing checkpoint file.
          content:
            application/zip:
              schema:
                $ref: '#/components/schemas/ZipFile'
          headers:
            Content-Disposition:
              description: Indicates that the response should be downloaded as an attachment.
              schema:
                type: string
...
  /project/{idProject}/volume/{idVolume}/result/{idResult}/download:
    get:
      summary: Download result files
      operationId: downloadResult
      tags:
        - Result
      parameters:
        - $ref: '#/components/parameters/idProject'
        - $ref: '#/components/parameters/idVolume'
        - $ref: '#/components/parameters/idResult'
      responses:
        '200':
          description: ZIP file containing result files.
          content:
            application/zip:
              schema:
                $ref: '#/components/schemas/ZipFile'
          headers:
            Content-Disposition:
              description: Indicates that the response should be downloaded as an attachment.
              schema:
                type: string
  /project/{idProject}/volume/{idVolume}/result/{idResult}/download/{fileIndex}:
    get:
      summary: Download a specific result file
      operationId: downloadResultFile
      tags:
        - Result
      parameters:
        - $ref: '#/components/parameters/idProject'
        - $ref: '#/components/parameters/idVolume'
        - $ref: '#/components/parameters/idResult'
        - name: fileIndex
          in: path
          required: true
          schema:
              type: integer
      responses:
        '200':
          description: ZIP file containing a result file.
          content:
            application/zip:
              schema:
                $ref: '#/components/schemas/ZipFile'
          headers:
            Content-Disposition:
              description: Indicates that the response should be downloaded as an attachment.
              schema:
                type: string
...
components:
  schemas:
    Project:
      $ref: "./schemas/Project.yaml"
    Volume:
      $ref: "./schemas/Volume.yaml#/Volume"
    RawData:
      $ref: "./schemas/RawData.yaml"
    SparseLabelVolumeData:
      $ref: "./schemas/SparseLabelVolumeData.yaml"
    PseudoLabelVolumeData:
      $ref: "./schemas/PseudoLabelVolumeData.yaml"
    Model:
      $ref: "./schemas/Model.yaml"
...
    Error:
      $ref: "./schemas/Error.yaml"
    ZipFile:
      type: string
      format: binary
      description: The binary data of the ZIP file.
      example: download.zip
  parameters:
    idProject:
      name: idProject
      in: path
      required: true
      schema:
          type: integer
    idVolume:
      name: idVolume
      in: path
      required: true
      schema:
          type: integer
    idVolumeData:
      name: idVolumeData
      in: path
      required: true
      schema:
          type: integer
    idModel:
      name: idModel
      in: path
      required: true
      schema:
          type: integer
    idCheckpoint:
      name: idCheckpoint
      in: path
      required: true
      schema:
          type: integer
    idResult:
      name: idResult
      in: path
      required: true
      schema:
          type: integer
  securitySchemes:
    express-session:
      type: apiKey
      in: cookie
      name: connect.sid